# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Exercise:
    properties:
      created_at:
        type: string
      description:
        type: string
      difficulty:
        description: beginner, intermediate, advanced
        type: string
      equipment:
        type: string
      id:
        type: integer
      image_url:
        type: string
      instructions:
        type: string
      muscle_group:
        type: string
      name:
        type: string
      training_exercises:
        description: Relationships
        items:
          $ref: '#/definitions/models.TrainingExercise'
        type: array
      updated_at:
        type: string
      video_url:
        type: string
    type: object
...
  models.Professor:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      specialties:
        type: string
      trainings:
        description: Relationships
        items:
          $ref: '#/definitions/models.Training'
        type: array
      updated_at:
        type: string
    type: object
...
  models.Training:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      professor:
        $ref: '#/definitions/models.Professor'
      professor_id:
        type: integer
      repetitions:
        type: integer
      sets:
        type: integer
      training_exercises:
        items:
          $ref: '#/definitions/models.TrainingExercise'
        type: array
      training_registrations:
        items:
          $ref: '#/definitions/models.TrainingRegistration'
        type: array
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: integer
    type: object
  models.TrainingCreate:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      professor_id:
        type: integer
      repetitions:
        type: integer
      sets:
        type: integer
    required:
    - name
    - professor_id
    type: object
  models.TrainingExercise:
    properties:
      exercise:
        $ref: '#/definitions/models.Exercise'
      exercise_id:
        type: integer
      id:
        type: integer
      repetitions:
        type: integer
      sets:
        type: integer
      training:
        allOf:
        - $ref: '#/definitions/models.Training'
        description: Relationships
      training_id:
        type: integer
    type: object
  models.TrainingExerciseCreate:
    properties:
      exercise_id:
        type: integer
      repetitions:
        type: integer
      sets:
        type: integer
    required:
    - exercise_id
    type: object
  models.TrainingRegistration:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      training:
        $ref: '#/definitions/models.Training'
      training_id:
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: integer
    type: object
  models.TrainingUpdate:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      professor_id:
        type: integer
      repetitions:
        type: integer
      sets:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      training_registrations:
        items:
          $ref: '#/definitions/models.TrainingRegistration'
        type: array
      trainings:
        description: Relationships
        items:
          $ref: '#/definitions/models.Training'
        type: array
      updated_at:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive workout training application with user management and
    exercise tracking
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SmartForce API
  version: "1.0"
paths:
...
  /trainings:
    get:
      consumes:
      - application/json
      description: Get all trainings with pagination
      parameters:
      - default: 0
        description: Skip records
        in: query
        name: skip
        type: integer
      - default: 100
        description: Limit records
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Training'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all trainings
      tags:
      - trainings
    post:
      consumes:
      - application/json
      description: Create a new training
      parameters:
      - description: Training data
        in: body
        name: training
        required: true
        schema:
          $ref: '#/definitions/models.TrainingCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Training'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new training
      tags:
      - trainings
  /trainings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a training
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete training
      tags:
      - trainings
    get:
      consumes:
      - application/json
      description: Get a specific training by ID
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Training'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get training by ID
      tags:
      - trainings
    put:
      consumes:
      - application/json
      description: Update training information
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      - description: Training data
        in: body
        name: training
        required: true
        schema:
          $ref: '#/definitions/models.TrainingUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Training'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update training
      tags:
      - trainings
  /trainings/{id}/exercises:
    post:
      consumes:
      - application/json
      description: Add an exercise to a training
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      - description: Exercise data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.TrainingExerciseCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TrainingExercise'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add exercise to training
      tags:
      - trainings
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"